# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/modul/{id}/update/{upload_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed update upload info as JSON (not part of TUS protocol)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Modul Update"
                ],
                "summary": "Get update upload information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Module ID being updated",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Upload ID",
                        "name": "upload_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancel an ongoing modul update upload and clean up resources",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Modul Update"
                ],
                "summary": "Cancel update upload (TUS DELETE)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Module ID being updated",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Upload ID",
                        "name": "upload_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content (upload cancelled)",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "Tus-Resumable": {
                                "type": "string",
                                "description": "TUS protocol version"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid upload ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Upload not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Upload already completed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "head": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Check the current progress of a modul update upload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Modul Update"
                ],
                "summary": "Get update upload status (TUS HEAD)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Module ID being updated",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Upload ID",
                        "name": "upload_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "TUS headers with current offset and length",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "Tus-Resumable": {
                                "type": "string",
                                "description": "TUS protocol version"
                            },
                            "Upload-Length": {
                                "type": "string",
                                "description": "Total file length"
                            },
                            "Upload-Offset": {
                                "type": "string",
                                "description": "Current offset in bytes"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid upload ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Upload not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload a chunk of file data for updating an existing module",
                "consumes": [
                    "application/offset+octet-stream"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Modul Update"
                ],
                "summary": "Upload update chunk (TUS PATCH)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Module ID being updated",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Upload ID from InitiateModulUpdateUpload",
                        "name": "upload_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "TUS protocol version (1.0.0)",
                        "name": "Tus-Resumable",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Current offset in bytes",
                        "name": "Upload-Offset",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Must be application/offset+octet-stream",
                        "name": "Content-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chunk size in bytes",
                        "name": "Content-Length",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content (upload continuing)",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "Tus-Resumable": {
                                "type": "string",
                                "description": "TUS protocol version"
                            },
                            "Upload-Offset": {
                                "type": "string",
                                "description": "New offset after this chunk"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid chunk or offset",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Upload not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Offset mismatch",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "412": {
                        "description": "TUS version mismatch",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "413": {
                        "description": "Chunk too large",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "415": {
                        "description": "Invalid content type",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/modul/{id}/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Start a new TUS upload for updating an existing module file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Modul Update"
                ],
                "summary": "Initiate modul update upload (TUS POST)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Module ID to update",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "TUS protocol version (1.0.0)",
                        "name": "Tus-Resumable",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Total file size in bytes",
                        "name": "Upload-Length",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Base64-encoded metadata (filename, etc.)",
                        "name": "Upload-Metadata",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "TUS error response",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "TUS error response",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Module not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "412": {
                        "description": "TUS version mismatch",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "413": {
                        "description": "File too large",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "Upload queue full",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "domain.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
    }
}